`update` alert dialog delete
